remove unused construct - restructure
